move data blood value to data package
